Add basic end-to-end tests for esologs server.
The change imports a few logs (2020-12-29 to 2021-01-03) into a
testdata directory, and sets up golden files of the expected outputs
of the logs server pages. It also adds the minimal `cpp-httplib` as an
external dependency, to provide a HTTP client to allow test code to
interact with the server.

Finally, there's a handful of minor reorganization to both the esologs
server and the common civetweb wrapper, to allow the tests to spawn an
instance of the server in an ephemeral port, and figure out what port
got used.