Refactor log server for multi-network; break stalker mode.
Caution: This commit entirely breaks the stalker mode. While both the
server and the bot framework are now multi-network-capable, the logger
plugin is still per-bot-instance. I need to figure out an incremental
upgrade path from the current setup, and whether to use separate
per-target stalker sockets or to overhaul the whole end-to-end stalker
protocol (both bot/server and server/browser) to be natively
multi-target. In the latter case, one alternative is to take this
chance and switch to using the brpc "worker" model for the bot/server
communication too. Decisions, decisions.