Reimplement stalker mode.
This commit also flips the existing scheme upside-down: the server
socket is now owned by the bot rather than the web server, and it is
the web server process that connects a client to it. It's pretty
arbitrary, but this feels a bit more natural.

The connection between the web server and the bot is now called a
"pipe", to avoid some confusion between the (websocket) connection
between the browser and the web server (which is still called
"stalker").