Avoid splitting log pipe events.
Previously there was a complicated on-stack copy if it happened that a
proto was split by the ring buffer wraparound. While this probably
never happens in practice (as long as the buffer keeps getting
consumed at a reasonable rate), let's still push that complexity into
the already existing ring buffer method that ensures contiguity of the
next allocation.