Add raw IRC logging support.
This commit adds a configuration option that can tell the bot to
capture everything[*] received or sent over the IRC
connection(s). This is useful as a safeguard against implementation
errors in the main per-target logging path: if it filters out
something it shouldn't, or if it turns out the server is sending some
useful information to include, those can be later backfilled to the
presentation logs.

Eventually, the raw logging might also provide the basis for
supporting redundant loggers connected from different systems, with
asynchronous unification. But that might be overengineering it.

[*] Almost; data sent by the low-level IRC connection handling code
isn't included in the logs, as it's never passed to the module hooks
so doesn't reach the logging module either.