Improve logging logic to avoid double-logging some messages.
Previously even the message types that are echoed back from the server
to the client were being logged. The bot doesn't really send very
many, but this would have triggered, e.g., if bottool was used to set
a whimsical quit message. The commit makes only PRIVMSG and NOTICE to
be logged also when outgoing, since all other messages are echoed
back.