Fix stalker websocket handler.
Looks like CivetWeb now wants the expected subprotocol to be
registered ahead of time in order for it to register. We could
probably just use the default protocol, TBH, but it is now the way it
is.

The protocol list also needs to be in unmovable memory for as long as
the server is alive, which I learned the hard way...