Add `ETag`/`Last-Modified` support to `logs.esolangs.org`.
Not bothering for the index pages, but for the log content ones, now
setting `ETag` and `Last-Modified` in the response, and respecting
`If-None-Match` and `If-Modified-Since` in the request.

This is very bare-bones support, but should be enough for common
crawling use-cases.

The `ETag` header of any page that's guaranteed to no longer get any
writes is set to the constant string "frozen", and its `Last-Modified`
timestamp will be set to five minutes after the last nominally
possible write (just to make sure there's no going back in time near
the day boundary).

For pages that are still accumulating data, the `ETag` header is set
to "N-S", where N and S are decimal numbers indicating:

- N: the day of the month. For a single-day log page, this is the
  corresponding day. For a monthly page, it's the day that's currently
  receiving data.
- S: the size of the matching logfile. Since files are only compressed
  with a few days of delay, this acts as a monotonically increasing
  counter.

The `Last-Modified` header of an active page is set to the file
modification timestamp of the logfile that's receiving data.

As a special case, for a monthly page where there's no existing
logfile of the current day actively being written (but the month
itself is not yet in the past):

- The `ETag` header is synthesized as "N-0", where N is set to the day
  that any new messages would be written to.
- The `Last-Modified` header is synthesized to indicate the (starting)
  midnight of that same day.