Use the `hinnant_date` library for date formatting.
Turns out GCC 12 on Debian stable (where this program will end up
running on and is convenient to be built on) doesn't know about the
`<format>` header.

We're already using the date library for parsing, might as well stick
to it for formatting as well. Five years down the road it'll probably
be possible to switch to native C++20 (or later) support and drop the
external dependency entirely.