Update to Gemma 2.0 and latest gemma.cpp version.
The 2.0 version of the 2B model size was released, so let's just swap
to that. It also requires a version bump for gemma.cpp, and a few
changes to the APIs.

Also importing some build changes done to make it work on the Clang
version on Debian stable (lower minimum CMake version, explicitly
enable C++20 and import <string_view>).